# docs/swagger.yaml
definitions:
  errs.ValidationError:
    properties:
      message:
        type: string
    type: object
...
  params.CreateTagRequest:
    properties:
      names:
        items:
          type: string
        type: array
    type: object
...
  params.GetTagResponse:
    properties:
      last_used:
        type: string
      name:
        type: string
      trending_score:
        type: number
      usage_count:
        type: integer
    type: object
  params.LoginUserRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  params.RegisterUserRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
    type: object
...
  rest.APIError:
    properties:
      error:
        type: string
    type: object
info:
  contact:
    email: mrezaelange@gmail.com
    name: reza
  description: 'JWT Authorization header using the Bearer scheme. Example: "Authorization:
    Bearer {token}"'
  title: Content Management System API
  version: "1.0"
paths:
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate a user with email and password.
      parameters:
      - description: Login User Request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/params.LoginUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errs.ValidationError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest.APIError'
      summary: Login User
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user with the provided details.
      parameters:
      - description: Register User Request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/params.RegisterUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: ok
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errs.ValidationError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest.APIError'
      summary: Register User
      tags:
      - Auth
...
  /tags:
    get:
      consumes:
      - application/json
      description: Retrieve a list of tags with optional sorting.
      parameters:
      - description: fill with Bearer token. The token can be accessed via api /auth/login.
          For accessing draft, published, and archived articles. If authorization
          is not provided default is showing published articles
        in: header
        name: Authorization
        type: string
      - description: Sort by usage_count:asc | usage_count:desc | trending_score:asc
          | trending_score:desc | name:asc | name:desc | last_used:asc | last_used:desc
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/params.GetTagResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errs.ValidationError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest.APIError'
      security:
      - BearerAuth: []
      summary: Get Tags
      tags:
      - Tags
    post:
      consumes:
      - application/json
      description: Create a new tag with the provided names.
      parameters:
      - description: fill with Bearer token. The token can be accessed via api /auth/login.
          For accessing draft, published, and archived articles. If authorization
          is not provided default is showing published articles
        in: header
        name: Authorization
        type: string
      - description: Create Tag Request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/params.CreateTagRequest'
      produces:
      - application/json
      responses:
        "201":
          description: created
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errs.ValidationError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest.APIError'
      security:
      - BearerAuth: []
      summary: Create Tag
      tags:
      - Tags
  /tags/{name}:
    get:
      consumes:
      - application/json
      description: Retrieve a specific tag by name.
      parameters:
      - description: fill with Bearer token. The token can be accessed via api /auth/login.
          For accessing draft, published, and archived articles. If authorization
          is not provided default is showing published articles
        in: header
        name: Authorization
        type: string
      - description: Tag name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/params.GetTagResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errs.ValidationError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest.APIError'
      security:
      - BearerAuth: []
      summary: Get Tag
      tags:
      - Tags
swagger: "2.0"